telemetry: use uint8_t instead of int for stats_key bitfields

+++ telemetry/stats_key.h
@@ -1,19 +1,19 @@
 // used as the key in the BPF hashmap
 struct stats_key {
-    unsigned int af:1;
-    unsigned int qr_bit:1;
-    unsigned int do_bit:1;
-    unsigned int ad_bit:1;
-    unsigned int rrl_triggered:1;
+    uint8_t af:1;
+    uint8_t qr_bit:1;
+    uint8_t do_bit:1;
+    uint8_t ad_bit:1;
+    uint8_t rrl_triggered:1;
 
-    unsigned int no_edns:1;
-    unsigned int edns_size_leq1231:1;
-    unsigned int edns_size_1232:1;
-    unsigned int edns_size_leq1399:1;
-    unsigned int edns_size_1400:1;
-    unsigned int edns_size_leq1499:1;
-    unsigned int edns_size_1500:1;
-    unsigned int edns_size_gt1500:1;
+    uint8_t no_edns:1;
+    uint8_t edns_size_leq1231:1;
+    uint8_t edns_size_1232:1;
+    uint8_t edns_size_leq1399:1;
+    uint8_t edns_size_1400:1;
+    uint8_t edns_size_leq1499:1;
+    uint8_t edns_size_1500:1;
+    uint8_t edns_size_gt1500:1;
 
     uint8_t qtype;
     char tld[10];
